Finish fixing new migrations instace issues core

+++ backend/core/src/main/java/decidish/com/core/model/recipes/RecipeIngredient.java
@@ -22,6 +22,6 @@ public class RecipeIngredient {
     @JoinColumn(name = "ingredient_id")
     private Ingredient ingredient;
 
-    private Double quantity;
+    private BigDecimal quantity;
     private String unit;
 }
